+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.9.4
-- `LineSegmentExtensions.Intersect(this IList<Line> lines)`
+- `LineSegmentExtensions.Intersections(this IList<Line> lines)`
 - `Elements.Search.BinaryTree<T>`
 - `Elements.Search.BinaryTreeNode<T>`